renamed primary executable to gnip_search.py search.py and search_api.py are symlinks to gnip_search.py updated setup.py to install search.py, search_api.py, and gnip_search.py fixed default "timeline" behavior to produce JSON output
