+++ setup.py
@@ -6,7 +6,7 @@ setup(
     author='Scott Hendrickson, Josh Montague',
     author_email='scott@drskippy.net',
     packages=['gnip_search'],
-    scripts=['search_api.py','search.py'],
+    scripts=['gnip_search.py','search_api.py','search.py'],
     url='https://github.com/DrSkippy27/Gnip-Python-Search-API-Utilities',
     download_url='https://github.com/DrSkippy27/Gnip-Python-Search-API-Utilities/tags/',
     license='LICENSE.txt',
